Clean up a bit more
